traffic sign detection
In this project, I am sharing my work on the detection and recognition of traffic signs. This project, developed for a competition, is based on the utilization of the YOLOv5 model. It includes sample videos and images of the competition track. The project encompasses innovative methods and algorithms for accurately detecting and distinguishing traffic signs. Our system is designed with the aim of contributing to traffic safety. By examining our project, you can acquire detailed information on the detection of traffic signs.